adaptive trapz integral

+++ y2011union/data.py
@@ -6,7 +6,7 @@ import numpy as np
 path_to_data = os.path.dirname(os.path.abspath(__file__)) + '/raw-data/'
 distances_file = pd.read_csv(path_to_data + 'distances.txt', sep = ' ')
 selected_columns = distances_file[['z', 'mu', 'sigma_mu']]
-covariance_matrix = np.loadtxt(path_to_data + 'covariance_stat_sys.txt', delimiter='\t')
+covariance_matrix = np.loadtxt(path_to_data + 'covariance_stat_sys.txt', delimiter='\t', dtype=np.float64)
 
 z_values = np.array(selected_columns['z'], dtype = np.float64, copy = False)
 sort_indices = np.argsort(z_values)
